code: Завершен семантический анализ. Осталось заассемблировать NASMерть.

+++ Domain/Entities/Real.cs
@@ -5,7 +5,7 @@ namespace Domain.Entities
 {
     public class Real:Token
     {
-        public float Value { get; init; }
+        public double Value { get; init; }
         public Real(float value) : base(Tag.REAL_CONST, TokenType.CONST) => Value = value;
         public override string? ToString() => "real";
     }
